Move Neomodel class extensions to schemas.py
