+++ backend/database/models/officer.py
@@ -1,4 +1,4 @@
-from backend.database.neo_classes import JsonSerializable, PropertyEnum
+from backend.schemas import JsonSerializable, PropertyEnum
 
 from neomodel import (
     StructuredNode,
